+++ extensions/promptlab/src/hooks/useModel.ts
@@ -19,7 +19,7 @@ export default function useModel(
   input: string,
   temperature: string,
   execute: boolean,
-  modelOverride?: Model
+  modelOverride?: Model,
 ) {
   const preferences = getPreferenceValues<ExtensionPreferences>();
   const [data, setData] = useState<string>("");
@@ -50,7 +50,7 @@ export default function useModel(
     outputTiming: "async",
     lengthLimit: "2500",
     temperature: "1.0",
-    name: "Text-Davinci-003 Via Raycast AI",
+    name: "GPT-3.5 Turbo Instruct Via Raycast AI",
     description: "",
     favorited: false,
     id: "",
@@ -142,18 +142,18 @@ export default function useModel(
             "{prompt}",
             preferences.promptPrefix +
               prompt.replaceAll(/[\n\r\s]+/g, " ").replaceAll('"', '\\"') +
-              preferences.promptSuffix
+              preferences.promptSuffix,
           )
           .replace(
             "{basePrompt}",
-            preferences.promptPrefix + basePrompt.replaceAll(/[\n\r\s]+/g, " ").replaceAll('"', '\\"')
+            preferences.promptPrefix + basePrompt.replaceAll(/[\n\r\s]+/g, " ").replaceAll('"', '\\"'),
           )
           .replace(
             "{input}",
             targetModel.inputSchema.includes("{prompt") && prompt == input
               ? ""
-              : input.replaceAll(/[\n\r\s]+/g, " ").replaceAll('"', '\\"') + preferences.promptSuffix
-          )
+              : input.replaceAll(/[\n\r\s]+/g, " ").replaceAll('"', '\\"') + preferences.promptSuffix,
+          ),
       );
   if (preferences.includeTemperature) {
     modelSchema["temperature"] = temp;
@@ -173,17 +173,17 @@ export default function useModel(
             "{prompt}",
             preferences.promptPrefix +
               prompt.replaceAll(/[\n\r\s]+/g, " ").replaceAll('"', '\\"') +
-              preferences.promptSuffix
+              preferences.promptSuffix,
           )
           .replace(
             "{basePrompt}",
-            preferences.promptPrefix + basePrompt.replaceAll(/[\n\r\s]+/g, " ").replaceAll('"', '\\"')
+            preferences.promptPrefix + basePrompt.replaceAll(/[\n\r\s]+/g, " ").replaceAll('"', '\\"'),
           )
           .replace(
             "{input}",
             targetModel.inputSchema.includes("{prompt") && prompt == input
               ? ""
-              : input.replaceAll(/[\n\r\s]+/g, " ").replaceAll('"', '\\"') + preferences.promptSuffix
+              : input.replaceAll(/[\n\r\s]+/g, " ").replaceAll('"', '\\"') + preferences.promptSuffix,
           ),
       });
       const tag = basePrompt + prompt + input;
@@ -279,7 +279,7 @@ export default function useModel(
         ...useAI(filterString(preferences.promptPrefix + prompt + preferences.promptSuffix, 5000), {
           execute: execute,
           creativity: temp,
-          model: targetModel.endpoint == "Raycast AI 3.5" ? "gpt-3.5-turbo" : "text-davinci-003",
+          model: targetModel.endpoint == "Raycast AI 3.5" ? "gpt-3.5-turbo" : "gpt-3.5-turbo-instruct",
         }),
         dataTag: basePrompt + prompt + input,
         stopModel: stopModel,