+++ extensions/promptlab/src/utils/command-utils.ts
@@ -2,8 +2,10 @@ import { objcImports, replaceAllHandler, rselectHandler, splitHandler, trimHandl
 import { exec } from "child_process";
 import { Command, CommandOptions, StoreCommand } from "./types";
 import { LocalStorage, AI } from "@raycast/api";
-import { Placeholders } from "./placeholders";
 import { runAppleScript } from "@raycast/utils";
+import { PLApplicator } from "placeholders-toolkit";
+import { PromptLabPlaceholders, loadCustomPlaceholders } from "../lib/placeholders";
+import { loadAdvancedSettingsSync } from "./storage-utils";
 
 /**
  * Runs the action script of a PromptLab command, providing the AI response as the `response` variable.
@@ -24,12 +26,16 @@ export const runActionScript = async (
   prompt: string,
   input: string,
   response: string,
-  type?: string
+  type?: string,
 ) => {
+  const settings = loadAdvancedSettingsSync();
+  const customPlaceholders = await loadCustomPlaceholders(settings);
+
   try {
     if (type == "applescript" || type == undefined) {
       await runAppleScript(
-        await Placeholders.bulkApply(`${objcImports}
+        await PLApplicator.bulkApply(
+          `${objcImports}
       ${splitHandler}
       ${trimHandler}
       ${replaceAllHandler}
@@ -37,7 +43,12 @@ export const runActionScript = async (
       set prompt to "${prompt.replaceAll('"', '\\"')}"
       set input to "${input.replaceAll('"', '\\"')}"
       set response to "${response.replaceAll('"', '\\"')}"
-      ${script}`)
+      ${script}`,
+          {
+            customPlaceholders,
+            defaultPlaceholders: PromptLabPlaceholders,
+          },
+        ),
       );
     } else if (type == "zsh") {
       const runScript = (script: string): Promise<string> => {
@@ -47,13 +58,18 @@ export const runActionScript = async (
         ${script.replaceAll("\n", " && ")}`;
 
         return new Promise((resolve, reject) => {
-          Placeholders.bulkApply(shellScript).then((subbedScript) => {
-            exec(subbedScript, (error, stdout) => {
-              if (error) {
-                reject(error);
-                return;
-              }
-              resolve(stdout);
+          loadCustomPlaceholders(settings).then((customPlaceholders) => {
+            PLApplicator.bulkApply(shellScript, {
+              customPlaceholders,
+              defaultPlaceholders: PromptLabPlaceholders,
+            }).then((subbedScript) => {
+              exec(subbedScript, (error, stdout) => {
+                if (error) {
+                  reject(error);
+                  return;
+                }
+                resolve(stdout);
+              });
             });
           });
         });
@@ -93,9 +109,11 @@ export const runReplacements = async (
   prompt: string,
   context: { [key: string]: string },
   disallowedCommands: string[],
-  options?: CommandOptions
+  options?: CommandOptions,
 ): Promise<string> => {
   let subbedPrompt = prompt;
+  const settings = loadAdvancedSettingsSync();
+  const customPlaceholders = await loadCustomPlaceholders(settings);
 
   // Replace config placeholders
   if (options != undefined && options.setupConfig != undefined) {
@@ -108,7 +126,11 @@ export const runReplacements = async (
     }
   }
 
-  subbedPrompt = await Placeholders.bulkApply(subbedPrompt, context);
+  subbedPrompt = await PLApplicator.bulkApply(subbedPrompt, {
+    customPlaceholders,
+    defaultPlaceholders: PromptLabPlaceholders,
+    context,
+  });
 
   // Replace command placeholders
   for (const cmdString of Object.values(await LocalStorage.allItems())) {
@@ -118,7 +140,7 @@ export const runReplacements = async (
       (subbedPrompt.includes(`{{${cmd.name}}}`) || subbedPrompt.includes(`{{${cmd.id}}}`))
     ) {
       const cmdResponse = await AI.ask(
-        await runReplacements(cmd.prompt, context, [cmd.name, cmd.id, ...disallowedCommands])
+        await runReplacements(cmd.prompt, context, [cmd.name, cmd.id, ...disallowedCommands]),
       );
       if (cmd.actionScript != undefined && cmd.actionScript.trim().length > 0 && cmd.actionScript != "None") {
         await runActionScript(cmd.actionScript, cmd.prompt, "", cmdResponse, cmd.scriptKind);
@@ -140,7 +162,7 @@ export const runReplacements = async (
 export const updateCommand = async (
   oldCommandData: Command | undefined,
   newCommandData: Command,
-  setCommands?: React.Dispatch<React.SetStateAction<Command[]>>
+  setCommands?: React.Dispatch<React.SetStateAction<Command[]>>,
 ) => {
   const commandData = await LocalStorage.allItems();
   const commandDataFiltered = Object.values(commandData).filter((cmd, index) => {
@@ -152,7 +174,7 @@ export const updateCommand = async (
   });
 
   if (setCommands != undefined) {
-    setCommands([...commandDataFiltered?.map((data) => JSON.parse(data)), newCommandData]);
+    setCommands([...commandDataFiltered.map((data) => JSON.parse(data)), newCommandData]);
   }
 
   if (oldCommandData != undefined && oldCommandData.name != newCommandData.name) {