+++ extensions/promptlab/src/utils/types.ts
@@ -444,7 +444,7 @@ export type ModelManager = {
   createModel: (
     newData: Model & {
       [key: string]: string | boolean;
-    }
+    },
   ) => Promise<false | Model>;
 
   /**