+++ extensions/visual-studio-code-recent-projects/src/index.tsx
@@ -185,7 +185,7 @@ function LocalItem(props: { entry: EntryLike; uri: string; pinned?: boolean } & 
                 shortcut={{ modifiers: ["cmd", "shift"], key: "o" }}
                 onAction={() =>
                   open(path, terminalApp).catch(() =>
-                    showToast(Toast.Style.Failure, `Failed to open with ${terminalApp.name}`)
+                    showToast(Toast.Style.Failure, `Failed to open with ${terminalApp?.name}`)
                   )
                 }
               />