+++ extensions/linear/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Linear Changelog
 
+## [Add issue links] - 2023-12-29
+
+- Add support for issue links. It is now possible to see the number of links an issue has from the detail view. You can also browse them in a dedicated view.
+
 ## [Mark notification as read without opening it] - 2023-11-16
 
 - You can now mark notification as read in the menu bar without opening the notification by pressing `⌥` and clicking the notification.